feat!: remove title from Popup::new
BREAKING CHANGE: The `Popup::new` method no longer accepts a title
parameter. Instead, the title can be set using the `title` method on the
`Popup` instance.

```diff
- let popup = Popup::new("tui-popup demo", "Press any key to exit")
+ let popup = Popup::new("Press any key to exit")
+     .title("tui-popup demo")
      .style(Style::new().white().on_blue());
```
